pyproject.toml
[tool.poetry]
name = "avito_skypro"
version = "0.1.0"
description = ""
authors = ["neony <neonyan15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
Pillow = "^9.1.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
